+++ examples/low_level_api/low_level_api_chat_cpp.py
@@ -76,8 +76,14 @@ specified) expect poor results""",
         self.lparams.n_parts = self.params.n_parts
         self.lparams.seed = self.params.seed
         self.lparams.memory_f16 = self.params.memory_f16
-        self.lparams.use_mlock = self.params.use_mlock
-        self.lparams.use_mmap = self.params.use_mmap
+        if self.params.use_mmap and self.params.use_mlock:
+            self.lparams.load_mode = llama_cpp.LLAMA_LOAD_MODE_MMAP_MLOCK
+        elif self.params.use_mlock:
+            self.lparams.load_mode = llama_cpp.LLAMA_LOAD_MODE_MLOCK
+        elif self.params.use_mmap:
+            self.lparams.load_mode = llama_cpp.LLAMA_LOAD_MODE_MMAP
+        else:
+            self.lparams.load_mode = llama_cpp.LLAMA_LOAD_MODE_NONE
 
         self.model = llama_cpp.llama_model_load_from_file(
             self.params.model.encode("utf8"), self.lparams